Создан TripController
